+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Which is based on [Keep A Changelog](http://keepachangelog.com/)
 
 - Prefix error message with ERROR 
 - Fix new line with clear input
+- Replace sha1 to sha512 in pbkdf2.Key function
 
 ## v1.0.0 - 2019-07-12